refactor: extract recorded and public device action flows (#1079)
Phase 4 of the layout.svelte.ts split (#910). Splits the recorded
command flows out of layout/command-adapters.ts into focused domain
modules and moves the multi-step public device flows out of the facade:

- layout/recorded-device-type-actions.ts: add/update/delete device type
  and bulk delete recorded actions
- layout/recorded-device-actions.ts: place/move/remove and per-property
  device update recorded actions
- layout/recorded-rack-actions.ts: rack update, batch update, and clear
  rack recorded actions
- layout/device-actions.ts: duplicateDevice, placeInContainer, and
  moveDeviceToRack public flows
- layout/command-adapters.ts now contains only the command store adapter

The facade keeps the public getLayoutStore() API unchanged and injects
$state.snapshot via DI where extracted flows need plain-object clones
(duplicateDevice, moveDeviceToRack), mirroring removeDeviceRecorded.

Closes #1079
