+++ src/lib/stores/layout/recorded-device-type-actions.ts
@@ -1,0 +1,202 @@
+/**
+ * Recorded Device Type Actions for Layout Store
+ *
+ * Extracted from layout/command-adapters.ts — device type library
+ * operations with undo/redo support. Each function creates a Command
+ * wrapping raw mutators, then executes it through the history system.
+ */
+
+import type { Cable, DeviceType, PlacedDevice } from "$lib/types";
+import {
+  createDeviceType as createDeviceTypeHelper,
+  findDeviceType as findDeviceTypeInArray,
+  type CreateDeviceTypeInput,
+} from "$lib/stores/layout-helpers";
+import { layoutDebug } from "$lib/utils/debug";
+import { getHistoryStore } from "../history.svelte";
+import {
+  createAddDeviceTypeCommand,
+  createUpdateDeviceTypeCommand,
+  createDeleteDeviceTypeCommand,
+  createBatchCommand,
+} from "../commands";
+import type { LayoutStateAccess } from "./types";
+import { getCommandStoreAdapter } from "./command-adapters";
+import { getPlacedDevicesWithRackForType } from "./mutators";
+
+/**
+ * Add a device type with undo/redo support
+ * @param ctx - Layout state access
+ * @param data - Device type creation input
+ * @returns The created device type
+ */
+export function addDeviceTypeRecorded(
+  ctx: LayoutStateAccess,
+  data: CreateDeviceTypeInput,
+): DeviceType {
+  const deviceType = createDeviceTypeHelper(data);
+  const history = getHistoryStore();
+  const adapter = getCommandStoreAdapter(ctx);
+
+  const command = createAddDeviceTypeCommand(deviceType, adapter);
+  history.execute(command);
+  ctx.markDirty();
+
+  return deviceType;
+}
+
+/**
+ * Update a device type with undo/redo support
+ * @param ctx - Layout state access
+ * @param slug - Device type slug
+ * @param updates - Properties to update
+ */
+export function updateDeviceTypeRecorded(
+  ctx: LayoutStateAccess,
+  slug: string,
+  updates: Partial<DeviceType>,
+): void {
+  const layout = ctx.getLayout();
+  const existing = findDeviceTypeInArray(layout.device_types, slug);
+  if (!existing) return;
+
+  // Capture before state for the fields being updated
+  const before: Partial<DeviceType> = {};
+  for (const key of Object.keys(updates) as (keyof DeviceType)[]) {
+    before[key] = existing[key] as never;
+  }
+
+  const history = getHistoryStore();
+  const adapter = getCommandStoreAdapter(ctx);
+
+  const command = createUpdateDeviceTypeCommand(slug, before, updates, adapter);
+  history.execute(command);
+  ctx.markDirty();
+}
+
+/**
+ * Find cables connected to any of the given placed devices.
+ * Used so DELETE_DEVICE_TYPE can clean up dangling cable endpoints (#1483).
+ */
+function findCablesForDevices(
+  ctx: LayoutStateAccess,
+  placedDevices: { rackId: string; device: PlacedDevice }[],
+): Cable[] {
+  const layout = ctx.getLayout();
+  const cables = layout.cables;
+  if (!cables || cables.length === 0) return [];
+  const deviceIds = new Set(placedDevices.map((p) => p.device.id));
+  if (deviceIds.size === 0) return [];
+  return cables.filter(
+    (c) => deviceIds.has(c.a_device_id) || deviceIds.has(c.b_device_id),
+  );
+}
+
+/**
+ * Delete a device type with undo/redo support
+ * @param ctx - Layout state access
+ * @param slug - Device type slug
+ */
+export function deleteDeviceTypeRecorded(
+  ctx: LayoutStateAccess,
+  slug: string,
+): void {
+  const layout = ctx.getLayout();
+  const existing = findDeviceTypeInArray(layout.device_types, slug);
+  if (!existing) return;
+
+  const placedDevices = getPlacedDevicesWithRackForType(ctx, slug);
+  const connectedCables = findCablesForDevices(ctx, placedDevices);
+  const history = getHistoryStore();
+  const adapter = getCommandStoreAdapter(ctx);
+
+  const command = createDeleteDeviceTypeCommand(
+    existing,
+    placedDevices,
+    adapter,
+    connectedCables,
+  );
+  history.execute(command);
+  ctx.markDirty();
+}
+
+/**
+ * Delete multiple device types with single undo/redo support
+ * Used for bulk cleanup operations
+ * @param ctx - Layout state access
+ * @param slugs - Array of device type slugs to delete
+ * @returns Number of device types actually deleted
+ */
+export function deleteMultipleDeviceTypesRecorded(
+  ctx: LayoutStateAccess,
+  slugs: string[],
+): number {
+  layoutDebug.state(
+    "deleteMultipleDeviceTypesRecorded: received %d slugs",
+    slugs.length,
+  );
+
+  if (slugs.length === 0) {
+    layoutDebug.state(
+      "deleteMultipleDeviceTypesRecorded: early return - no slugs",
+    );
+    return 0;
+  }
+
+  const layout = ctx.getLayout();
+  const history = getHistoryStore();
+  const adapter = getCommandStoreAdapter(ctx);
+  const commands: ReturnType<typeof createDeleteDeviceTypeCommand>[] = [];
+  // A cable connecting devices of two different types would otherwise be
+  // snapshotted by both per-type delete commands, restoring it twice on undo.
+  const claimedCableIds = new Set<string>();
+
+  for (const slug of slugs) {
+    const existing = findDeviceTypeInArray(layout.device_types, slug);
+    if (!existing) continue;
+
+    const placedDevices = getPlacedDevicesWithRackForType(ctx, slug);
+    const connectedCables = findCablesForDevices(ctx, placedDevices).filter(
+      (cable) => {
+        if (claimedCableIds.has(cable.id)) return false;
+        claimedCableIds.add(cable.id);
+        return true;
+      },
+    );
+    const command = createDeleteDeviceTypeCommand(
+      existing,
+      placedDevices,
+      adapter,
+      connectedCables,
+    );
+    commands.push(command);
+  }
+
+  if (commands.length === 0) {
+    layoutDebug.state(
+      "deleteMultipleDeviceTypesRecorded: no valid commands created",
+    );
+    return 0;
+  }
+
+  // Create a batch command for single undo
+  const count = commands.length;
+  const description =
+    count === 1 ? "Delete device type" : `Delete ${count} device types`;
+
+  layoutDebug.state(
+    "deleteMultipleDeviceTypesRecorded: executing batch command - %s",
+    description,
+  );
+
+  const batchCommand = createBatchCommand(description, commands);
+  history.execute(batchCommand);
+  ctx.markDirty();
+
+  layoutDebug.state(
+    "deleteMultipleDeviceTypesRecorded: completed - deleted %d device types",
+    count,
+  );
+
+  return count;
+}
